Issue_5  Add support for r10k Puppetfile's opts
Add feature for Librarian to recognise r10k's :commit, :tag, and :branch
as accepted aliases for Librarian's :ref. Librarian will fail if more
than one of these options are specified.

A known issue with this implementation is that, internally, all of these
options are still treated as git refs. In other words, Librarian will
not notice if the user calls what is really a tag a commit and so on.