Be specific about exposing git environment variables (#9)
The regex used to output the git environment variables only checks the
the start of the variable begins with `GIT`. In most circumstances, this
would be fine. However, if you have environment variables for `GITHUB`
(say access tokens) these are also exposed and this becomes a potential
security risk.

To remedy this, I've just updated the regex to check the environment
variable name begins with `GIT_` which based on my research _should_
cover all the core variables that can be expected here.